poker_ai_mvp/requirements.txt
# Core Dependencies
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.2.0
opencv-python>=4.8.1
Pillow>=10.0.1
numpy>=1.24.3
pandas>=2.1.1
paddleocr>=2.7.3
mss>=9.0.1

# Database & Data Processing
sqlalchemy>=2.0.21

# GUI & Visualization  
matplotlib>=3.7.2

# Configuration & Logging
pydantic>=2.4.2
python-dotenv>=1.0.0
loguru>=0.7.2
PyYAML>=6.0

# Async & Performance
psutil>=5.9.0
pywin32>=306

# Development & Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.1